Use -D_FILE_OFFSET_BITS=64 for off_t

+++ storage.c
@@ -146,21 +146,21 @@ int sc_storage_gc_run(struct systemc *sc)
 	return reclaimed;
 }
 
-int sc_storage_get_free(struct systemc *sc)
+off_t sc_storage_get_free(struct systemc *sc)
 {
-	int fs_free, fs_min;
+	off_t fs_free, fs_min;
 	struct statfs buf;
 
 	if (statfs("/storage/trails/0.json", &buf) < 0)
 		return -1;
 
-	fs_free = (int) buf.f_bsize * (int) buf.f_bfree;
-	fs_min = (int) buf.f_bsize * (int) buf.f_blocks;
+	fs_free = (off_t) buf.f_bsize * (off_t) buf.f_bfree;
+	fs_min = (off_t) buf.f_bsize * (off_t) buf.f_blocks;
 
 	// Always leave 5%
 	fs_min -= (fs_min * 95) / 100;
 
-	sc_log(DEBUG, "fs_free: %d, fs_min: %d", fs_free, fs_min); 
+	sc_log(DEBUG, "fs_free: %llu, fs_min: %llu", fs_free, fs_min);
 
 	if (fs_free < fs_min)
 		return 0;
